Main_Application/requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
langchain-chroma
python-dotenv
fitz
pymupdf
starlette